pyproject.toml
[tool.poetry]
name = "rag-mcp"
version = "0.1.0"
description = "RAG-MCP"
authors = ["lukeding <dingkai9627@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.119.0"
uvicorn = {extras = ["standard"], version = "^0.37.0"}
aiosqlite = "^0.19.0"
python-multipart = "^0.0.20"
aiofiles = "^23.2.1"
httpx = "^0.28.1"
python-docx = "^1.2.0"
pypdf2 = "^3.0.1"
python-magic = "^0.4.27"
python-dotenv = "^1.1.1"
openai = "^2.3.0"
loguru = "^0.7.3"
reportlab = "^4.4.4"
mcp = "^1.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
